// src/prosemirror/schema.d.ts
import { Schema } from 'prosemirror-model';
declare const schema: Schema<"aside" | "blockquote" | "button" | "cite" | "iframe" | "image" | "switch" | "text" | "display" | "range" | "doc" | "paragraph" | "heading" | "code_block" | "horizontal_rule" | "hard_break" | "ordered_list" | "bullet_list" | "list_item" | "callout" | "cite_group" | "math" | "equation" | "variable" | "dynamic", "abbr" | "code" | "em" | "link" | "strong" | "underline" | "superscript" | "subscript" | "strikethrough">;
export default schema;